DeploymentConfig health update
